ci: add get-branch and get-tag jobs
Get Branch step is used to get the current branch name and use it on Get
Tag step, this step is used to set the npm tag when deploying the
package.

By using this new job we can still maintain older versions of the
package, instead of always realising the latest updates.

For example: this package is currently on version 1.2.0 and we have a
service that uses the 0.48.x version. In that service I cannot update for 1.2.0
cause it has breaking changes, like nodejs runtime and packages imports.
So to update/fix something on this package I will need, also, to upgrade
my hole service to fit the latest changes on this package.